+++ src/comments/comments.module.ts
@@ -3,9 +3,10 @@ import { CommentsController } from './comments.controller';
 import { CommentsService } from './comments.service';
 import { PrismaModule } from '../prisma/prisma.module';
 import { SceneAccessService } from '../workspace/scene-access.service';
+import { NotificationsModule } from '../notifications/notifications.module';
 
 @Module({
-  imports: [PrismaModule],
+  imports: [PrismaModule, NotificationsModule],
   controllers: [CommentsController],
   providers: [CommentsService, SceneAccessService],
   exports: [CommentsService],